+++ backend/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for queries without source data in vector database
 - Graceful failure of triple export when no chunks are found
 
+### Changed
+
+- Separated embedding service from LLM service
+
 ## [v0.1.5] - 2024-10-29
 
 ### Changed